fix tests for task item ui
